[FSDE-30731] Fix PR comments
