+++ src/documentation/pages/templates-wip/card/index.vue
@@ -1,15 +1,12 @@
 <template lang="pug">
   div
     <TitleBar title="Card" description="A collection of templates built with responsive Cards." />
-    .chi-grid__container.-pt--3
-        <MarketingFeatureCards />
-        <CategoryCards />
+    <Examples />
 </template>
 
 <script lang="ts">
 import { Component, Vue } from 'vue-property-decorator';
-import MarketingFeatureCards from './_marketing-feature-cards.vue';
-import CategoryCards from './_category-cards.vue';
+import Examples from './examples/index.vue';
 
 @Component({
   head() {
@@ -18,8 +15,7 @@ import CategoryCards from './_category-cards.vue';
     }
   },
   components: {
-    MarketingFeatureCards,
-    CategoryCards
+    Examples
   },
   layout() {
     return 'wide';
