+++ src/documentation/pages/templates-wip/error-404/examples/index.vue
@@ -1,0 +1,16 @@
+<template lang="pug">
+  .chi-grid__container.-pt--3
+    <Error />
+</template>
+
+<script lang="ts">
+import { Component, Vue } from 'vue-property-decorator';
+import Error from './_error.vue';
+
+@Component({
+  components: {
+    Error
+  }
+})
+export default class Examples extends Vue {}
+</script>
